creation and usage of asset_group metadata.py file
